# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Team-Ops API
  version: 0.1.0
  description: Self-hosted engineering board for humans and AI agents.
servers:
  - url: /v1
...
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      responses:
        "200":
          $ref: "#/components/responses/OrganizationList"
    post:
      tags: [Organizations]
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          $ref: "#/components/responses/Organization"
  /organizations/{orgId}/members:
    get:
      tags: [Organizations]
      operationId: listMembers
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          $ref: "#/components/responses/MemberList"
    post:
      tags: [Organizations]
      operationId: addMember
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRequest"
      responses:
        "201":
          $ref: "#/components/responses/Member"
  /organizations/{orgId}/invites:
    get:
      tags: [Organizations]
      operationId: listInvites
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          $ref: "#/components/responses/InviteList"
    post:
      tags: [Organizations]
      operationId: createInvite
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "201":
          $ref: "#/components/responses/Invite"
  /organizations/{orgId}/invites/{inviteId}:
    delete:
      tags: [Organizations]
      operationId: revokeInvite
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - in: path
          name: inviteId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
components:
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Envelope:
      type: object
      required: [data, error, meta]
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/ErrorBody"
        meta:
          type: object
    ErrorBody:
      type: object
      nullable: true
      properties:
        code:
          type: string
        message:
          type: string
...
    CreateOrganizationRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        color: { type: string }
...
    MemberRequest:
      type: object
      required: [email, role]
      properties:
        email: { type: string }
        role: { type: string, enum: [admin, developer, viewer] }
...
    CreateInviteRequest:
      type: object
      properties:
        email: { type: string }
        role: { type: string, enum: [admin, developer, viewer] }
...
  responses:
...
    OK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
    Organization:
      description: Organization
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    OrganizationList:
      description: Organizations
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Member:
      description: Member
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    MemberList:
      description: Members
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Invite:
      description: Invite
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    InviteList:
      description: Invites
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...